cli: Move Git help parser support
The CLI parser builds subcommands and owns the Git help parser class that
routes --help through packaged or system man pages.

The project does not provide a separate home for Git help lookup, so parser
construction carries manpage resource lookup, MANPATH construction, and git
help invocation next to command registration. That makes the parsing module
harder to scan as more command branches are added.

This commit addresses that by moving Git help parsing and manpage lookup
into a dedicated CLI module while keeping argument_parser.py focused on
command parser assembly.

The following tests commit will move the Git help tests and add an
architecture guard for the new boundary.
